Type check Self type

+++ Reef.Core/TypeChecking/Expressions/ValueAccessor.cs
@@ -35,7 +35,7 @@ public partial class TypeChecker
             var signature = GetUnionSignature(DefId.Result);
             var instantiatedUnion = InstantiateUnion(signature, [], boxed ? Token.Boxed(SourceSpan.Default) : Token.Unboxed(SourceSpan.Default), SourceRange.Default);
 
-            var okVariant = GetUnionVariant(instantiatedUnion, variantName)
+            var okVariant = GetUnionVariant(instantiatedUnion, variantName, UInt64())
                             ?? throw new UnreachableException($"{variantName} is a built in variant of Result");
 
             if (okVariant is not TupleUnionVariant { BoxedCreateFunction: var createFunction, UnboxedCreateFunction: var unboxedCreateFunction })
